# perf(duckdb): fix dataframe insertion performance